Move file TestCaseFilterExpression.cs from CrossPlat to common
